% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bam2Bsseq.R
\name{BSseq2}
\alias{BSseq2}
\title{Construct BSseq objects from BED files}
\usage{
BSseq2(
  M = NULL,
  Cov = NULL,
  filter = NULL,
  coef = NULL,
  se.coef = NULL,
  trans = NULL,
  parameters = NULL,
  pData = NULL,
  gr = NULL,
  pos = NULL,
  chr = NULL,
  sampleNames = NULL,
  rmZeroCov = FALSE
)
}
\arguments{
\item{M}{matrix, methylation data}

\item{Cov}{matrix, coverage data}

\item{filter}{matrix, ambiguous methylation modification data}

\item{coef}{matrix, smoothing estimates}

\item{se.coef}{matrix, smoothing standard errors}

\item{trans}{function, smoothing transformation}

\item{parameters}{list, smoothing parameters}

\item{pData}{data frame, phenotypic data with samples as rows and variables as columns}

\item{gr}{GRanges, genomic positions}

\item{pos}{vector, locations}

\item{chr}{vector, chromosomes}

\item{sampleNames}{vector, sample names}

\item{rmZeroCov}{logical, should genomic locations with zero coverage in all samples be removed}

\item{files}{vector, BED files.}
}
\value{
BSseq object
}
\description{
Construct BSseq objects from BED files
}
\details{
These functions read in nanopore data, modified BED files, to Bsseq objects
}
